Build the p2p feature in CI
Nothing built defra-node or db with p2p on, so a dep missing from the feature
list only broke downstream consumers. Fixes the clippy warning it surfaced.

+++ crates/p2p/src/sync/push_backlog.rs
@@ -531,7 +531,7 @@ impl PushBacklog {
         let jitter_bp = {
             let mut hasher = std::collections::hash_map::DefaultHasher::new();
             std::hash::Hash::hash(&(&peer_key, consecutive), &mut hasher);
-            u64::from(std::hash::Hasher::finish(&hasher) % 500)
+            std::hash::Hasher::finish(&hasher) % 500
         };
         let cooldown = base + (base * jitter_bp as u32) / 1000;
         inner.peer_cooldowns.insert(
